+++ application/helpers/values_helper.php
@@ -59,52 +59,6 @@ if (! function_exists('decrypt'))
 	}
 }
 
-if (! function_exists('game_hash'))
-{
-	/**
-	 * Generate hashed password for game account
-	 *
-	 * @param string $username
-	 * @param string $password
-	 * @param string $type
-	 * @param mixed $salt
-	 * @return mixed
-	 */
-	function game_hash($username, $password, $type = 'classic', $salt = null)
-	{
-		switch ($type)
-		{
-			case 'classic':
-				return strtoupper(sha1(strtoupper($username) . ':' . strtoupper($password)));
-				break;
-			case 'bnet':
-				return strtoupper(bin2hex(strrev(hex2bin(strtoupper(hash('sha256', strtoupper(hash('sha256', strtoupper($username)) . ':' . strtoupper($password))))))));
-				break;
-			case 'srp6':
-				// Constants
-				$g = gmp_init(7);
-				$N = gmp_init('894B645E89E1535BBDAD5B8B290650530801B18EBFBF5E8FAB3C82872A3E9BB7', 16);
-				// Calculate first hash
-				$h1 = sha1(strtoupper($username.':'.$password), TRUE);
-				// Calculate second hash
-				$h2 = sha1($salt.$h1, TRUE);
-				// Convert to integer (little-endian)
-				$h2 = gmp_import($h2, 1, GMP_LSW_FIRST);
-				// g^h2 mod N
-				$verifier = gmp_powm($g, $h2, $N);
-				// Convert back to a byte array (little-endian)
-				$verifier = gmp_export($verifier, 1, GMP_LSW_FIRST);
-				// Pad to 32 bytes, remember that zeros go on the end in little-endian!
-				$verifier = str_pad($verifier, 32, chr(0), STR_PAD_RIGHT);
-				return $verifier;
-				break;
-			default:
-				return '';
-				break;
-		}
-	}
-}
-
 if (! function_exists('money_converter'))
 {
 	/**